Goin/requirements.txt
aiofiles==23.1.0
aiogram==2.22.2
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
django-request==1.6.3
djangorestframework==3.14.0
drf-yasg==1.21.7
exceptiongroup==1.1.3
frozenlist==1.4.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
magic-filter==1.0.11
multidict==6.0.4
packaging==23.1
phonenumbers==8.13.21
Pillow==10.0.1
psycopg2-binary==2.9.9
pydantic==2.3.0
pydantic_core==2.6.3
python-dateutil==2.8.2
python-decouple==3.8
python-telegram-bot==20.5
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.6
uvicorn==0.23.2
yarl==1.9.2
